Big changes!
Skipped an update, but I have separated the menu buttons on the top frame and the entire options menu into their own respective classes. This is just to practice OOP concepts, calling class variables, and just to see if I could! Also, if I plan on adding anymore features it should make this more simpler to do so.